Add variety implementation with complexity analysis.
