+++ Union-Find/UnionFind.playground/Sources/UnionFindWeightedQuickFind.swift
@@ -1,0 +1,59 @@
+import Foundation
+
+/// Quick-find algorithm may take ~MN steps
+/// to process M union commands on N objects
+public struct UnionFindWeightedQuickFind<T: Hashable> {
+    private var index = [T: Int]()
+    private var parent = [Int]()
+    private var size = [Int]()
+    
+    public init() {}
+    
+    public mutating func addSetWith(_ element: T) {
+        index[element] = parent.count
+        parent.append(parent.count)
+        size.append(1)
+    }
+    
+    private mutating func setByIndex(_ index: Int) -> Int {
+        return parent[index]
+    }
+    
+    public mutating func setOf(_ element: T) -> Int? {
+        if let indexOfElement = index[element] {
+            return setByIndex(indexOfElement)
+        } else {
+            return nil
+        }
+    }
+    
+    public mutating func unionSetsContaining(_ firstElement: T, and secondElement: T) {
+        if let firstSet = setOf(firstElement), let secondSet = setOf(secondElement) {
+            if firstSet != secondSet {
+                if size[firstSet] < size[secondSet] {
+                    for index in 0..<parent.count {
+                        if parent[index] == firstSet {
+                            parent[index] = secondSet
+                        }
+                    }
+                    size[secondSet] += size[firstSet]
+                } else {
+                    for index in 0..<parent.count {
+                        if parent[index] == secondSet {
+                            parent[index] = firstSet
+                        }
+                    }
+                    size[firstSet] += size[secondSet]
+                }
+            }
+        }
+    }
+    
+    public mutating func inSameSet(_ firstElement: T, and secondElement: T) -> Bool {
+        if let firstSet = setOf(firstElement), let secondSet = setOf(secondElement) {
+            return firstSet == secondSet
+        } else {
+            return false
+        }
+    }
+}
